Remove the ask-me dedicated page
